A função de adicionar obra está funcionando perfeitamente (sem contar no deploy)

+++ src/pages/add-obra.jsx
@@ -5,36 +5,94 @@ export default function AdicionarObra() {
   const [listaObra, setListaObra] = useState([]);
   const [titulo, setTitulo] = useState("");
   const [desc, setDesc] = useState("");
-  const [img_url, setImg_Url] = useState("");
   const [categoria, setCategoria] = useState("");
   const [editandoId, setEditandoId] = useState(null);
+  const [imagem, setImagem] = useState(null);
+  const [uploading, setUploading] = useState(false);
+  const [previewUrl, setPreviewUrl] = useState(null);
+
+
 
 
 //CREATE
   const addObra = async () => {
-    const novaObraDados = {
-      titulo: titulo,
-      desc: desc,
-      img_url: img_url,
-      categoria: categoria
-    }
-    const { data, error } = await supabase.from("ListaObras").insert([novaObraDados]).single();
+  if (!titulo || !desc || !categoria || !imagem) {
+    alert("Erro ao adicionar obra, preencha todos os campos");
+    return;
+  }
 
-    if (error){
-      console.log("Erro ao adicionar nova obra", error)
-    } else{
-      setListaObra((prev) => [...prev, data]);
-      setTitulo("");
-      setDesc("");
-      setImg_Url("");
-      setCategoria("");
+  const img_url = await uploadImagem();
+  if(!img_url) return;
+
+  const novaObraDados = {
+    titulo,
+    desc,
+    categoria,
+    img_url
+  };
+
+  const { data, error } = await supabase
+    .from("ListaObras")
+    .insert([novaObraDados])
+    .select()
+    .single();
+
+  if (error) {
+    alert("Erro ao salvar obra" + error.message);
+    return;
+  }
+
+  setListaObra((prev) => [data, ...prev]);
+  setTitulo("");
+  setDesc("");
+  setCategoria("");
+  setImagem(null);
+  setPreviewUrl(null);
+
+  alert("Obra adicionada com sucesso!");
+};
+
+
+const uploadImagem = async () => {
+  if (!imagem) {
+    alert("Selecione uma imagem.");
+    return null;
+  }
+
+  const gerarSlug = (texto) => {
+    return texto
+      .toLowerCase()
+      .normalize("NFD").replace(/[\u0300-\u036f]/g, "")
+      .replace(/[^a-z0-9]+/g, "-")
+      .replace(/(^-|-$)/g, "");
+  };
+
+  const extensao = imagem.name.split('.').pop();
+  const nomeArquivo = `${gerarSlug(titulo)}.${extensao}`;
+
+
+  const {error} = await supabase.storage
+    .from("Imagens")
+    .upload(nomeArquivo, imagem)
+
+    if(error) {
+      alert("Erro ao enviar imagem");
+      return null;
     }
-  }; 
+
+  const {data} = supabase.storage
+    .from("Imagens")
+    .getPublicUrl(nomeArquivo)
+
+  return data.publicUrl
+}
 
 
   //READ
   const fetchObras = async () => {
-    const {data, error} = await supabase.from("ListaObras").select("*") //sem categorias selecionadas
+    const {data, error} = await supabase
+    .from("ListaObras")
+    .select("*") //sem categorias selecionadas
     if (error) {
       console.log("Erro ao ler as obras.", error)
     } else {
@@ -48,13 +106,14 @@ export default function AdicionarObra() {
 
 
   //UPDATE
-  const editarObra = async (id, titulo, desc, img_url, categoria) => {
+  const editarObra = async (id, titulo, desc, categoria, img_url) => {
   const { data, error } = await supabase
     .from("ListaObras")
-    .update({ titulo, desc, img_url, categoria })
+    .update({ titulo, desc, categoria, img_url })
     .eq("id", id)
     .select()
     .single();
+
 
   if (error) {
     console.log("Erro ao alterar obra.", error);
@@ -67,18 +126,33 @@ export default function AdicionarObra() {
 };
 
 //DELETE
-const deletarObra = async (id) => {
-    const { data, error } = await supabase
-      .from("ListaObras")
-      .delete()
-      .eq("id", id);
+const deletarObra = async (id, img_url) => {
+  // pega o nome do arquivo da URL
+  const nomeArquivo = img_url.split("/").pop();
 
-    if (error) {
-      console.log("Erro ao excluir obra: ", error);
-    } else {
-      setListaObra((prev) => prev.filter((obra) => obra.id !== id));
-    }
-  };
+  // remove do bucket
+  const { error: erroStorage } = await supabase.storage
+    .from("Imagens")
+    .remove([nomeArquivo]);
+
+  if (erroStorage) {
+    console.log("Erro ao deletar imagem:", erroStorage);
+    return;
+  }
+
+  // remove do banco
+  const { error } = await supabase
+    .from("ListaObras")
+    .delete()
+    .eq("id", id);
+
+  if (error) {
+    console.log("Erro ao excluir obra: ", error);
+  } else {
+    setListaObra((prev) => prev.filter((obra) => obra.id !== id));
+  }
+};
+
 
 
   
@@ -87,83 +161,94 @@ const deletarObra = async (id) => {
     <>
     <h1>Adicionar Obra</h1>
     <>
-    <input type="text" placeholder="Titulo:" onChange={(e) => setTitulo(e.target.value)}/>
-    <input type="text" placeholder="Descrição:" onChange={(e) => setDesc(e.target.value)}/>
-    <input type="text" placeholder="Imagem:" onChange={(e) => setImg_Url(e.target.value)}/>
-    <input type="text" placeholder="Categoria:" onChange={(e) => setCategoria(e.target.value)}/>
+    <input type="text" value={titulo} placeholder="Titulo:" onChange={(e) => setTitulo(e.target.value)}/>
+    <input type="text" value={desc} placeholder="Descrição:" onChange={(e) => setDesc(e.target.value)}/>
+    <input type="text" value={categoria} placeholder="Categoria:" onChange={(e) => setCategoria(e.target.value)}/>
+
+    <input 
+      type="file" 
+      accept="image/png, image/jpeg, image/jpg" 
+      onChange={(e) => {
+        const file = e.target.files[0];
+        setImagem(file);
+        setPreviewUrl(URL.createObjectURL(file));
+      }}
+    />
+
+    {previewUrl && <img src={previewUrl} width="150" />}
+
     <button onClick={addObra}>Adicionar obra</button>
     </>
 
     <>
     <ul>
-  {listaObra.map((obra) => (
-    <li key={obra.id} style={{ marginBottom: "20px" }}>
-      {editandoId === obra.id ? (
-        // MODO EDIÇÃO
-        <>
-          <input
-            value={obra.titulo}
-            onChange={(e) =>
-              setListaObra((prev) =>
-                prev.map((item) =>
-                  item.id === obra.id
-                    ? { ...item, titulo: e.target.value }
-                    : item
+    {listaObra.map((obra) => (
+      <li key={obra.id} style={{ marginBottom: "20px" }}>
+        {editandoId === obra.id ? (
+          // MODO EDIÇÃO
+          <>
+            <input
+              value={obra.titulo}
+              onChange={(e) =>
+                setListaObra((prev) =>
+                  prev.map((item) =>
+                    item.id === obra.id
+                      ? { ...item, titulo: e.target.value }
+                      : item
+                  )
                 )
-              )
-            }
-          />
-
-          <input
-            value={obra.desc}
-            onChange={(e) =>
-              setListaObra((prev) =>
-                prev.map((item) =>
-                  item.id === obra.id
-                    ? { ...item, desc: e.target.value }
-                    : item
+              }
+            />
+            <input
+              value={obra.desc}
+              onChange={(e) =>
+                setListaObra((prev) =>
+                  prev.map((item) =>
+                    item.id === obra.id
+                      ? { ...item, desc: e.target.value }
+                      : item
+                  )
                 )
-              )
-            }
-          />
-
-          <input
-            value={obra.img_url}
-            onChange={(e) =>
-              setListaObra((prev) =>
-                prev.map((item) =>
-                  item.id === obra.id
-                    ? { ...item, img_url: e.target.value }
-                    : item
+              }
+            />
+            <input
+              value={obra.img_url}
+              onChange={(e) =>
+                setListaObra((prev) =>
+                  prev.map((item) =>
+                    item.id === obra.id
+                      ? { ...item, img_url: e.target.value }
+                      : item
+                  )
                 )
-              )
-            }
-          />
-
-          <input
-            value={obra.categoria}
-            onChange={(e) =>
-              setListaObra((prev) =>
-                prev.map((item) =>
-                  item.id === obra.id
-                    ? { ...item, categoria: e.target.value }
-                    : item
+              }
+            />
+            <input
+              value={obra.categoria}
+              onChange={(e) =>
+                setListaObra((prev) =>
+                  prev.map((item) =>
+                    item.id === obra.id
+                      ? { ...item, categoria: e.target.value }
+                      : item
+                  )
                 )
-              )
-            }
-          />
-
+              }
+            />
           <button
-            onClick={() =>
-              editarObra(
-                obra.id,
-                obra.titulo,
-                obra.desc,
-                obra.img_url,
-                obra.categoria
-              )
-            }
+            onClick={() => {
+                editarObra(
+                  obra.id,
+                  obra.titulo,
+                  obra.desc,
+                  obra.img_url,
+                  obra.categoria
+                )
+
+                alert(`Obra salva com sucesso!`)
+              }}
           >
+            
             Salvar
           </button>
 
@@ -177,12 +262,12 @@ const deletarObra = async (id) => {
           <h1>{obra.titulo}</h1>
           <h2>{obra.categoria}</h2>
           <h3>{obra.desc}</h3>
-          <p>{obra.img_url}</p>
+          <img src={obra.img_url} width="200" />
 
           <button onClick={() => setEditandoId(obra.id)}>
             Editar
           </button>
-          <button onClick={() => deletarObra(obra.id)}>
+          <button onClick={() => deletarObra(obra.id, obra.img_url)}>
             Excluir
           </button>
         </>
